timy-backend/pyproject.toml
[project]
name = "timy-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "firebase-admin>=6.5.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "fireo>=2.2.2",
]
